test: Mock 설정 통합
